New reasonableness test based on acceleration and disregard unreasonable last positions
Reasonableness test on acceleration for at least 3 previous reasonable
positions and for distance between messages of at least 200 m and less
than 30 s (jitter can be normal). Previous speed of aircraft is weighted
since the calculated speed between only two positions is constantly
fluctuating.

Also, when a position turns out to be unreasonable, it will not be saved
as last_pos for the decoding of the next position message.